Implement 2_6
Implement initialize queue method and output method removing the first element from the queue. Add logic to print the queue and to insert elements at the end of the queue. Create a basic test script in main that tests the methods by running 2 for loops and prints out the results.